comu/com.energyict.mdc.device.lifecycle/src/main/java/com/energyict/mdc/device/lifecycle/impl/micro/checks/AllIssuesAreClosed.java
/*
 * Copyright (c) 2017 by Honeywell International Inc. All Rights Reserved
 */
package com.energyict.mdc.device.lifecycle.impl.micro.checks;

import com.elster.jupiter.fsm.State;
import com.energyict.mdc.device.data.Device;
import com.energyict.mdc.device.lifecycle.ExecutableMicroCheckViolation;
import com.energyict.mdc.device.lifecycle.config.DefaultTransition;
import com.energyict.mdc.device.lifecycle.config.MicroCategory;

import java.time.Instant;
import java.util.EnumSet;
import java.util.Optional;
import java.util.Set;

/**
 * Checks that all Issue's on a device are closed.
 */
public class AllIssuesAreClosed extends TranslatableServerMicroCheck {

    @Override
    public String getCategory() {
        return MicroCategory.ISSUES.name();
    }

    @Override
    public Optional<ExecutableMicroCheckViolation> execute(Device device, Instant effectiveTimestamp, State toState) {
        return device.hasOpenIssues() ?
                fail(MicroCheckTranslations.Message.ALL_ISSUES_AND_ALARMS_ARE_CLOSED) :
                Optional.empty();
    }

    @Override
    public Set<DefaultTransition> getOptionalDefaultTransitions() {
        return EnumSet.of(
                DefaultTransition.DEACTIVATE_AND_DECOMMISSION,
                DefaultTransition.DECOMMISSION);
    }
}
